requirements.txt
Django>=4.0,<5.0
psycopg2-binary
python-dotenv
django-extensions
Werkzeug
pyOpenSSL
django-environ
freezegun==1.2.2
python-dateutil
celery>=5.0,<6.0
redis>=3.5.3
dj-database-url==2.1.0
whitenoise
pytz
gunicorn==20.1.0
django-widget-tweaks
django-celery-beat==2.5.0
tzdata==2024.1
django-celery-results
backports.zoneinfo;python_version<"3.9"